// src/app/Models/PersonalProposal/amulyaloan.ts
export class Amulya {

    ID: any;
    CLIENT_ID: number;
    PROPOSAL_ID: number;
    
    NAME_OF_GROUP: string;
    ADDRESS_OF_GROUP: string;
    SAVING_AMOUNT: number;
    ANNUAL_INCOME: number;
    FATHER_NAME: string;
    ARCHIVE_FLAG:string;
    constructor(){
        this.ARCHIVE_FLAG = 'F'
    }
}